+++ SpaceFlight/Physics/Calculator/DragCalculator.cs
@@ -5,9 +5,9 @@ namespace SpaceFlight.Physics.Calculator
 {
     public class DragCalculator
     {
-        public static Force CalculateDrag(Angle dragAngle, double cd, double p, Speed speed, double area)
+        public static Force CalculateDrag(Angle dragAngle, decimal cd, decimal p, Speed speed, decimal area)
         {
-            var force = cd * 0.5 * p * Math.Pow(speed.Value, 2) * area;
+            var force = cd * 0.5m * p * (decimal) Math.Pow((double) speed.Value, 2) * area;
 
             return new Force(dragAngle, force);
         }
