+++ SpaceFlight/Physics/Calculator/GravityCalculator.cs
@@ -5,9 +5,9 @@ namespace SpaceFlight.Physics.Calculator
 {
     class GravityCalculator
     {
-        public static readonly double GravitationalConstant = 6.67330E-11;
+        public static readonly decimal GravitationalConstant = 6.67330E-11m;
 
         public static Force CalculateGravity(Mass p1, Mass p2, double r, Angle angle) =>
-            new Force(angle, GravitationalConstant * (p1.Value * p2.Value) / Math.Pow(r, 2));
+            new Force(angle, GravitationalConstant * (p1.Value * p2.Value) / (decimal) Math.Pow(r, 2));
     }
 }
